Move snowbridge transfer flow to useSnowbridgeSdk
