+++ app/src/hooks/useTransfer.tsx
@@ -16,10 +16,11 @@ import useNotification from './useNotification'
 import useOngoingTransfers from './useOngoingTransfers'
 import { getRoute } from '@/config/registry'
 import useSnowbridgeContext from './useSnowbridgeContext'
+import useSnowbridgeSdk from './useSnowbridgeSdk'
 
 export type Sender = JsonRpcSigner | SubstrateAccount
 
-interface TransferParams {
+export interface TransferParams {
   environment: Environment
   sender: Sender
   sourceChain: Chain
@@ -33,33 +34,11 @@ interface TransferParams {
 
 export type Status = 'Idle' | 'Loading' | 'Validating' | 'Sending'
 
-type ValidationResult = toEthereum.SendValidationResult | toPolkadot.SendValidationResult
-
 const useTransfer = () => {
   const [status, setStatus] = useState<Status>('Idle')
-  const { addTransfer: addTransferToStorage } = useOngoingTransfers()
-  const { addNotification } = useNotification()
-  const { snowbridgeContext } = useSnowbridgeContext()
+  const snowbridgeSdk = useSnowbridgeSdk()
 
-  const handleValidationFailure = (plan: ValidationResult) => {
-    console.error('Validation failed:', plan)
-    const errorMessage =
-      plan.failure && plan.failure.errors.length > 0
-        ? plan.failure.errors[0].message
-        : 'Transfer validation failed'
-    addNotification({ message: errorMessage, severity: NotificationSeverity.Error })
-  }
-
-  const handleSendError = (e: unknown) => {
-    console.error('Transfer error:', e)
-    if (!(e instanceof Error) || !e.message.includes('ethers-user-denied')) captureException(e)
-    addNotification({
-      message: 'Failed to submit the transfer',
-      severity: NotificationSeverity.Error,
-    })
-  }
-
-  // main transfer function which is exposed to the components.
+  // The entry point function which is exposed to the components
   const transfer = async ({
     environment,
     sender,
@@ -81,67 +60,7 @@ const useTransfer = () => {
       }
 
       case 'SnowbridgeApi': {
-        SnowbridgeSdk.transfer({
-          environment,
-          sender,
-          sourceChain,
-          token,
-          destinationChain,
-          recipient,
-          amount,
-          fees,
-          onSuccess,
-        })
-        break
-      }
-    }
-  }
-
-  namespace SnowbridgeSdk {
-    // main transfer function which is exposed to the components.
-    export const transfer = async ({
-      environment,
-      sender,
-      sourceChain,
-      token,
-      destinationChain,
-      recipient,
-      amount,
-      fees,
-      onSuccess,
-    }: TransferParams) => {
-      setStatus('Loading')
-      try {
-        if (snowbridgeContext === undefined) {
-          addNotification({
-            message: 'Some nuts and bolts are not quite there yet',
-            severity: NotificationSeverity.Error,
-          })
-          return
-        }
-
-        const direction = resolveDirection(sourceChain, destinationChain)
-
-        const plan = await validate(
-          direction,
-          snowbridgeContext,
-          sender,
-          sourceChain,
-          token,
-          destinationChain,
-          recipient,
-          amount,
-        )
-
-        if (plan.failure) {
-          handleValidationFailure(plan)
-          return
-        }
-
-        await performTransfer(
-          snowbridgeContext,
-          sender,
-          plan,
+        snowbridgeSdk.transfer(
           {
             environment,
             sender,
@@ -153,151 +72,9 @@ const useTransfer = () => {
             fees,
             onSuccess,
           },
-          direction,
+          setStatus,
         )
-      } catch (e) {
-        console.error('Transfer initialization error:', e)
-        addNotification({
-          message: 'Transfer failed!',
-          severity: NotificationSeverity.Error,
-        })
-      } finally {
-        setStatus('Idle')
-      }
-    }
-
-    // Executes the transfer. Validation should happen before.
-    const performTransfer = async (
-      context: Context,
-      sender: Sender,
-      plan: ValidationResult,
-      params: TransferParams,
-      direction: Direction,
-    ) => {
-      const {
-        sourceChain,
-        token,
-        destinationChain,
-        recipient,
-        amount,
-        environment,
-        fees,
-        onSuccess,
-      } = params
-      try {
-        setStatus('Sending')
-        let sendResult
-
-        switch (direction) {
-          case Direction.ToPolkadot:
-            sendResult = await toPolkadot.send(
-              context,
-              sender as Signer,
-              plan as toPolkadot.SendValidationResult,
-            )
-            break
-
-          case Direction.ToEthereum:
-            sendResult = await toEthereum.send(
-              context,
-              sender as WalletOrKeypair,
-              plan as toEthereum.SendValidationResult,
-            )
-            break
-
-          default:
-            throw new Error('Unsupported flow')
-        }
-
-        if (sendResult.failure) throw new Error('Transfer failed')
-
-        onSuccess?.()
-        addNotification({
-          message: 'Transfer initiated. See below!',
-          severity: NotificationSeverity.Success,
-        })
-
-        const senderAddress =
-          sender instanceof JsonRpcSigner
-            ? await sender.getAddress()
-            : (sender as WalletOrKeypair).address
-
-        const tokenData = await getErc20TokenUSDValue(token.address)
-        const tokenUSDValue =
-          tokenData && Object.keys(tokenData).length > 0 ? tokenData[token.address]?.usd : 0
-        const date = new Date()
-
-        addTransferToStorage({
-          id: sendResult.success!.messageId ?? 'todo', // TODO(nuno): replace with actual messageId
-          sourceChain,
-          token,
-          tokenUSDValue,
-          sender: senderAddress,
-          destChain: destinationChain,
-          amount: amount.toString(),
-          recipient,
-          date,
-          environment,
-          sendResult,
-          fees,
-        } satisfies StoredTransfer)
-
-        // metrics
-        if (environment === Environment.Mainnet) {
-          trackTransferMetrics({
-            sender: senderAddress,
-            sourceChain: sourceChain.name,
-            token: token.name,
-            amount: amount.toString(),
-            destinationChain: destinationChain.name,
-            usdValue: tokenUSDValue,
-            usdFees: fees.inDollars,
-            recipient: recipient,
-            date: date.toISOString(),
-          })
-        }
-      } catch (e) {
-        handleSendError(e)
-      } finally {
-        setStatus('Idle')
-      }
-    }
-
-    const validate = async (
-      direction: Direction,
-      context: Context,
-      sender: Sender,
-      sourceChain: Chain,
-      token: Token,
-      destinationChain: Chain,
-      recipient: string,
-      amount: bigint,
-    ): Promise<ValidationResult> => {
-      setStatus('Validating')
-      switch (direction) {
-        case Direction.ToPolkadot:
-          return await toPolkadot.validateSend(
-            context,
-            sender as Signer,
-            recipient,
-            token.address,
-            destinationChain.chainId,
-            amount,
-            BigInt(destinationChain.destinationFeeDOT || 0),
-          )
-
-        case Direction.ToEthereum:
-          return await toEthereum.validateSend(
-            context,
-            sender as WalletOrKeypair,
-            sourceChain.chainId,
-            recipient,
-            token.address,
-            amount,
-          )
-
-        default:
-          throw new Error('Unsupported flow')
+        break
       }
     }
   }
